+++ src/main/java/homeat/backend/domain/homeatcash/entity/WeekStatus.java
@@ -1,4 +1,4 @@
-package homeat.backend.domain.homeatcash;
+package homeat.backend.domain.homeatcash.entity;
 
 public enum WeekStatus {
     SUCCESS, FAIL, UNDO
